Improve Plugin Manager Cards to Match Dashboard Card Style
feat: add StatCard component for plugin statistics display in PluginManager

Fixes #2268
